Added files via upload
Anemia is a major pubic health problem affecting more than 500 million of the indian population and leading to 58 percent of maternal deaths in India. Currently many devices are available but needs a prick of blood and dedicated instruments to calculate the hemoglobin content accurately. Here we propose a non-invasive technique to periodically monitor hemoglobin content and can especially  be used to screen anemia on a regular basis in pregnant woman and children using pulse plethysmography. The device needs a single point calibration with zero subsequent screening cost. In addition device is portable and can be used in rural settings as a personal monitoring kit or can be used by a health worker.

This non-invasive device works on the principle of pulse plethysmography.
We have analysed the transmitted IR signal and found the standard deviation for 300 intervals, which is then calibrated against a known anemia detection technique. This one step calibration is then extrapolated in s